services/qa-runner/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
playwright==1.49.*
httpx==0.28.*
pydantic-settings==2.7.*
redis[hiredis]==5.2.*
Pillow==11.*
loguru==0.7.*
